Consolidate helpers into functions module
