--- constants.py
@@ -1,11 +1,0 @@
-AUTO_MM_SOLVENT_TRIGGER = "auto_detect_mm_solvent_from_input_xyz"
-TEMP_FRAME_XYZ_FILENAME = "_current_frame_data.xyz"
-LOG_FILENAME = "giqpy_run.log"
-
-JSON_KEY_NAME = "name"
-JSON_KEY_NATOMS = "nAtoms"
-JSON_KEY_CHARGE = "charge"
-JSON_KEY_SPIN_MULT = "spin_mult"
-JSON_KEY_MOL_FORMULA = "mol_formula"
-JSON_KEY_CHARGES_ARRAY = "charges"
-JSON_KEY_ELEMENT = "element"
